GC (#1496)
## Description

Adds the concept of GC.

There is an additional persisted table roots, which is name -> (hash,
format) where `name` is any blob. Currently `format` is either blob or
collection, but in the future this might become a multiformat or
something.

The GC computes the set of all live hashes from these roots and
user-provided extra roots. This is the mark phase, which mostly involves
IO for opening collections and the like.

Then it does a second phase which is the sweep phase. In this phase it
just looks at all blobs and deletes the ones which are not in the life
set. This involves the actual deletion, which may or may not be
expensive depending on the file system.

Anything hash that is added between the mark phase and the sweep phase
is protected against being deleted, and will only be considered during
the next mark phase.

To protect stuff against deletion there are two mechanisms:

- named roots, see above, which are persisted to disk
- temp pins, which are also (hash, format) pairs but are purely in
memory

While adding files or during sync, temp pins can be used to protect
temporary data from gc.

Todo:

- [x] actually persist the roots table
- [x] change RPC api to list named roots and to manipulate named roots
- [ ] think more about all the possible race conditions
- [x] format blob/collection vs multiformats format
- [x] document integration
- [x] add GC to the mem version

## Notes & open questions

All that I need from iroh-docs is an iterator of all hashes as (hash,
Format::Blob) pairs. ~The exact timing needs some thought for this to be
safe. I think it needs to happen after the life set is cleared, which
makes the current impl buggy.~

I added the ability to delete blobs even if they are pinned. Not sure if
everybody agrees, but I think you should be able to do this. Imagine you
got something that you really don't want to have due to legal reasons.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.