Move tickets around (#1897)
## Description

This moves the ticket code around.

The ticket trait goes into iroh-base and becomes public
Blob and node ticket go to iroh-net
Doc tickets remain in iroh itself

This also makes the trait completely open, kind is not an enum but a
`&'static str`. Postcard encoding is just encouraged via docs, not
enforced.

The purpose of this is to be able to use tickets such as the node ticket
from [dumbpipe](https://github.com/n0-computer/dumbpipe) and the blob
ticket from [sendme](https://github.com/n0-computer/sendme).

## Notes & open questions

~Versioning?~ => Added some very basic versioning/variant scheme - just
a single variant enum so the postcard bytes have a variant identifier.

I am currently reexporting the tickets from iroh-net in iroh, just so
you have all 3 ticket kinds in one place. Not sure if that is necessary.
Also, should the trait be called ticket and the individual impl also
called ticket?

This is valid in itself, but the way PublicKey seriaizes is broken. We
should merge https://github.com/n0-computer/iroh/pull/1900 first.

Note: yes, the additional version adds a byte. But the bugfix in #1900
removes one, so we're net zero bytes... :-)

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.