test(iroh-net): try fix flaky udp_blocked test
This test is flaky because somehow the executor is entirely blocked
for a solid 5 seconds after which the entire netcheck report is timing
out.  Thus the part of the netcheck report that is being tested never
executes.

What blocks the executor however is a mystery to me.  This fix assumes
that it is generating the backtrace in the log, because there simply
isn't anything else that it can be.  Thus the relevant change is no
longer producing the backtrace.  The error message is already pretty
obvious.

To  go further, this disables the anyhow backtrace feature entirely.
This will probably make our lifes a bit harder in some situations but
there are many places where they end up in the log otherwise and each
time this happens it blocks an executor thread for a long time.

This also tweaks the error reporting to not generate so much logging
noise when the captive portal fails in an expected way.

## Notes & open questions

An attempt at fixing #1901