refactor(iroh-bytes): Rewrite the blob store to use redb (#2051)
## Description

Rewrite the blob store to use redb

This does a number of things:
1. rewrite the blob store to use redb.
2. inline small data blobs as well as small outboards in the redb db
3. properly handle remotes lying about sizes
4. make sure all redb interactions, which can be IO, are done in an
actor on a thread

It also adds a consistency check of the store as a feature.

Since this is a totally new new store format, it provides functions to
import from a flat store, `import_flat_store` rather than having a
migration. This just takes the paths of some flat store directories and
adsorbs all the data in that location.

## Notes & open questions

- Open question: should the inline data and inline outboard store be
combined into one?
- Todo: audit the places where the actor terminates. If it does, the
store is dead. We don't want to do this unless there is something really
bad going on, on the other hand it does not make much sense to continue
with a broken store.

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.