fix(iroh-blobs): do not skip empty partial blobs in migration (#2604)
## Description

When a blob found during migration is missing an outboard, according to
the
`cli_bao_store_migration` test, we should still migrate it. Before this
change
the blob would be ignored. This makes the test no longer broken.

## Breaking Changes

n/a

## Notes & open questions

1. I can reproduce the failure 100% of the time, so before this change
**I consider the test broken, not flaky.**
2. I reproduced the failure by running an iroh node with the exact v0
data the test uses. After the change, I did this again, and also ran
both `validate` and `consistency-check` and got no errors.

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [ ] ~~Tests if relevant.~~
- [ ] ~~All breaking changes documented.~~