refactor(iroh-blobs, iroh)!: deprecate flat stores (#2629)
## Description

In vein of promoting doing things TheRightWay™, this PR deprecates flat
stores for their removal in a later iroh version. This is probably a low
stakes deprecation and removal but imo gets us in the flow of being
kind(er) with our users. Issue #2630 is created as a reminder to remove
this.

## Breaking Changes

- `iroh_blobs::store::fs::Store::import_flat_store` is deprecated.
Ensure all data is migrated before future versions, which won't support
this.
- `iroh_blobs::store::fs::FlatStorePaths` is deprecated. Ensure all data
is migrated before future versions, which won't support this.
- `iroh`: Flat stores are deprecated. Ensure all data is migrated before
future versions, which won't support this.

## Notes & open questions

n/a

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [ ] ~~Tests if relevant.~~
- [x] All breaking changes documented.