feat(iroh-net): allow customizing republish delay for the pkarr publisher (#2637)
## Description

The pkarr dht publisher has two delays, the initial delay and the
republish delay. These were constants so far, but now are configurable
in the builder.

This allows us to set the delay to 0s in the dht_discovery_smoke test.

## Breaking Changes

None

## Notes & open questions

Will this fix the flakiness?

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [x] All breaking changes documented.

Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>