refactor!: Migrate to tokio's AbortOnDropHandle (#2701)
## Description

This migrates us to tokio's AbortOnDropHandle from the various
versions we had for this ourselves.  Only for iroh::node::Node we keep
a local version around since it is both Clone *and* awaits the tasks.
Generally it's probably better to not make this kind of design though,
if possible.

## Breaking Changes

Removes the `iroh_net::util` module.  This only contained utilities
that should have been internal in the first place.  If you were using
one of the tools to abort tasks on drop, we recommend migrating to
`tokio_util::task::AbortOnDropHandle`.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
- [x] All breaking changes documented.