feat: simplify LocalPool handling
It has been a footgun for a few users, including myself, to make sure to keep around the `LocalPool`.
This changes the behaviour to construct a `LocalPool` and keep it around by default. If necessary, in the builder one can provide a custom handle, if there is the need for a custom pool.