feat!: simplify LocalPool handling (#47)
## Description

It has been a footgun for a few users, including myself, to make sure to
keep around the `LocalPool`. This changes the behaviour to construct a
`LocalPool` and keep it around by default. If necessary, in the builder
one can provide a custom handle, if there is the need for a custom pool.

## Breaking Changes

- remove `net_protocol::Blobs::new`, use the builder instead
- remove the `LocalPoolHandle` argument from
`net_protocol::Builder::build`