Move entire multiprovider downloader into an example.
This is a bit unfortunate, but there are major refactorings coming for the
store, and the current code has to fake a part of the functionality that
is needed from the store.

This way people can take a look at the code structure and also try it out.
Once the main store changes are merged, the multiprovider downloader will
move into the main crate.