feat: Richer tags api (#69)
## Description

This makes the anemic tags api much more rich

- Tag deletion
  - delete by range, by prefix, single item, or all 💣 
- Tag listing
  - list by range, prefix or get single item
- Tag setting
  - set a tag to a value
- Tag renaming
  - atomic rename, just like for files

## Breaking Changes

Breaks the RPC protocol and the store trait

RPC protocol:

- rpc::proto::tags::DeleteRequest now has from and to field
- rpc::proto::tags::DeleteRequest no longer has name field
- rpc::proto::tags::ListRequest now has from and to field
- rpc::proto::tags::ListRequest::all removed
- rpc::proto::tags::ListRequest::raw removed
- rpc::proto::tags::ListRequest::hash_seq removed

Store trait:

- store::traits::ReadableStore::tags now takes from and to

## Notes & open questions

Question: should we enable the rpc feature flag by default? Given that
the only way to get the pleasant API at this time is the rpc, and we
want beginners to have a good experience? In my big refactor you will be
able to get this API without rpc enabled, but for now this is just how
it is.

Implements https://github.com/n0-computer/iroh-blobs/issues/65

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.