feat: implement new blobs with bitfield tracking and observation (#92)
## Description

This replaces the old blobs with a new implementation that allows
tracking bitfields of incomplete blobs, as well as allows observing the
state of an incomplete blob.

## Breaking Changes

A lot.

## Notes & open questions

Note: this version is as of now considered alpha quality. People that
look for stability should use the 0.35 version of both iroh and
iroh-blobs for now.

## Change checklist

- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [ ] ~~All breaking changes documented.~~