Use single threaded version of tokio::sync::watch
Since we are guaranteed to be single threaded within the handler fns for a
single hash, we can use a much more lightweight watcher impl.

Also, tokio::sync::watch is optimized for the case where there are a lot
of parallel watchers, using a giant BigNotify internally. We want to
optimize for the case where there are just 0 or a few watchers, so we are
not that much concerned with the thundering herd problem.