refactor: simplify state of fs store entry (#102)
## Description

A few things were stored twice, just to enable Drop. Now we don't need
Drop anymore and can store them only once.

BaoFileHandle was a wrapper around a BaoFileHandleInner, which had the
actual BaoFileStorage as well as the additional info needed to be able
to implement Drop. Now we don't implement Drop anymore and can just pass
in the required extra info in the persist call from on_shutdown!

In several other places we now have to give additional info when working
with BaoFileHandle, but this is conveniently grouped in HashContext.

## Breaking Changes

None

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.