chore: Feature gate redb based store. (#143)
## Description

Feature gate redb based store. It is still enabled by default. Many
tests and examples require this to be enabled, but I think that is fine,
since we run both with all and with no default features.

## Breaking Changes

None

## Notes & open questions

None

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.