feat: Make node_id in ClientConnected optional (#148)
## Description

For some reason we can't get the node_id for 0rtt connections. Blobs
does not use 0rtt, but it might in the future. So we don't have the
node_id in all cases when we need to create a ClientConnected event.

See https://github.com/n0-computer/iroh/issues/3123

## Breaking Changes

No additional breaking changes compared to the last published version,
since the ClientConnected event is new.

## Notes & open questions

Question: should we add some more stuff to the ClientConnected event
since we might not have the node id? Socket addrs or something?

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.