feat: abstract over stream types on provide and get side (#147)
## Description

This PR introduces abstract versions of iroh::endpoint::{SendStream,
RecvStream} and modifies the provide side and get side implementation to
use these abstract streams instead of directly using iroh::endpoint
streams.

This is necessary for wrapping the streams into a transformation such as
compression, see the discussion in
https://github.com/n0-computer/sendme/pull/93 .

The compression example shows how streams can be wrapped into
compression/decompression to create a derived protocol with a different
ALPN that is identical to the blobs protocol except for compression.

## Breaking Changes

iroh::endpoint::SendStream and iroh::endpoint::RecvStream are replaced
with the traits iroh_blobs::util::SendStream and
iroh_blobs::util::RecvStream in the get FSM and in the provider side
API.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.

---------

Co-authored-by: Frando <frando@unbiskant.org>